Extracted parts of file utils string into yiniHelper functions
